fix analysis input location tests
