+++ sootup.java.bytecode/src/test/java/sootup/java/bytecode/inputlocation/PathBasedAnalysisInputLocationTest.java
@@ -39,6 +39,7 @@ import org.junit.Assert;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import sootup.core.frontend.BodySource;
+import sootup.core.inputlocation.AnalysisInputLocation;
 import sootup.core.inputlocation.EagerInputLocation;
 import sootup.core.model.*;
 import sootup.core.signatures.FieldSubSignature;
@@ -60,222 +61,6 @@ import sootup.java.core.views.JavaView;
  */
 @Category(Java8Test.class)
 public class PathBasedAnalysisInputLocationTest extends AnalysisInputLocationTest {
-
-  @Test
-  public void multiReleaseJar() {
-    final ClassType classType =
-        getIdentifierFactory().getClassType("de.upb.swt.multirelease.Utility");
-    final ClassType classType2 =
-        getIdentifierFactory().getClassType("de.upb.swt.multirelease.Main");
-
-    final JavaProject project_min =
-        JavaProject.builder(new JavaLanguage(Integer.MIN_VALUE))
-            .addInputLocation(PathBasedAnalysisInputLocation.create(mrj, null))
-            .build();
-    final JavaView view_min = project_min.createView();
-
-    final JavaProject project_8 =
-        JavaProject.builder(new JavaLanguage(8))
-            .addInputLocation(PathBasedAnalysisInputLocation.create(mrj, null))
-            .build();
-    final JavaView view_8 = project_8.createView();
-
-    final JavaProject project_9 =
-        JavaProject.builder(new JavaLanguage(9))
-            .addInputLocation(PathBasedAnalysisInputLocation.create(mrj, null))
-            .build();
-    final JavaView view_9 = project_9.createView();
-
-    final JavaProject project_10 =
-        JavaProject.builder(new JavaLanguage(10))
-            .addInputLocation(PathBasedAnalysisInputLocation.create(mrj, null))
-            .build();
-    final JavaView view_10 = project_10.createView();
-
-    final JavaProject project_max =
-        JavaProject.builder(new JavaLanguage(Integer.MAX_VALUE))
-            .addInputLocation(PathBasedAnalysisInputLocation.create(mrj, null))
-            .build();
-    final JavaView view_max = project_max.createView();
-
-    // for java10
-    Assert.assertEquals(
-        "/META-INF/versions/9/de/upb/swt/multirelease/Utility.class",
-        view_10.getClass(classType).get().getClassSource().getSourcePath().toString());
-    Assert.assertEquals(
-        "/de/upb/swt/multirelease/Main.class",
-        view_10.getClass(classType2).get().getClassSource().getSourcePath().toString());
-
-    // assert that method is correctly resolved
-    Assert.assertTrue(
-        view_10
-            .getClass(classType)
-            .get()
-            .getMethod(
-                getIdentifierFactory()
-                    .getMethodSubSignature(
-                        "printVersion",
-                        getIdentifierFactory().getType("void"),
-                        Collections.emptyList()))
-            .get()
-            .getBody()
-            .toString()
-            .contains("java 9"));
-
-    // for java 9
-    Assert.assertEquals(
-        "/META-INF/versions/9/de/upb/swt/multirelease/Utility.class",
-        view_9.getClass(classType).get().getClassSource().getSourcePath().toString());
-    Assert.assertEquals(
-        "/de/upb/swt/multirelease/Main.class",
-        view_9.getClass(classType2).get().getClassSource().getSourcePath().toString());
-
-    // for java 8
-    Assert.assertEquals(
-        "/de/upb/swt/multirelease/Utility.class",
-        view_8.getClass(classType).get().getClassSource().getSourcePath().toString());
-    Assert.assertEquals(
-        "/de/upb/swt/multirelease/Main.class",
-        view_8.getClass(classType2).get().getClassSource().getSourcePath().toString());
-    // assert that method is correctly resolved to base
-    Assert.assertTrue(
-        view_8
-            .getClass(classType)
-            .get()
-            .getMethod(
-                getIdentifierFactory()
-                    .getMethodSubSignature(
-                        "printVersion",
-                        getIdentifierFactory().getType("void"),
-                        Collections.emptyList()))
-            .get()
-            .getBody()
-            .toString()
-            .contains("java 8"));
-
-    // for max int
-    Assert.assertEquals(
-        "/META-INF/versions/9/de/upb/swt/multirelease/Utility.class",
-        view_max.getClass(classType).get().getClassSource().getSourcePath().toString());
-    Assert.assertEquals(
-        "/de/upb/swt/multirelease/Main.class",
-        view_max.getClass(classType2).get().getClassSource().getSourcePath().toString());
-
-    // for min int
-    Assert.assertEquals(
-        "/de/upb/swt/multirelease/Utility.class",
-        view_min.getClass(classType).get().getClassSource().getSourcePath().toString());
-    Assert.assertEquals(
-        "/de/upb/swt/multirelease/Main.class",
-        view_min.getClass(classType2).get().getClassSource().getSourcePath().toString());
-  }
-
-  @Test
-  public void modularMultiReleaseJar() {
-    final ClassType utilityNoModule =
-        getIdentifierFactory().getClassType("de.upb.swt.multirelease.Utility");
-
-    final ModuleJavaClassType utilityModule =
-        JavaModuleIdentifierFactory.getInstance()
-            .getClassType("de.upb.swt.multirelease/de.upb.swt.multirelease.Utility");
-
-    final ClassType classType2 =
-        getIdentifierFactory().getClassType("de.upb.swt.multirelease.Main");
-
-    final JavaProject project_8 =
-        JavaProject.builder(new JavaLanguage(8))
-            .addInputLocation(PathBasedAnalysisInputLocation.create(mmrj, null))
-            .build();
-    final JavaView view_8 = project_8.createView();
-
-    final JavaModuleProject project_9 =
-        (JavaModuleProject)
-            JavaModuleProject.builder(new JavaLanguage(9))
-                .enableModules()
-                .addInputLocation(
-                    (ModuleInfoAnalysisInputLocation)
-                        PathBasedAnalysisInputLocation.create(mmrj, null))
-                .build();
-
-    final JavaModuleView view_9 = project_9.createView();
-
-    ModuleSignature moduleSignature =
-        JavaModuleIdentifierFactory.getModuleSignature("de.upb.swt.multirelease");
-
-    Assert.assertEquals(Collections.singleton(moduleSignature), view_9.getNamedModules());
-
-    Assert.assertTrue(view_9.getModuleInfo(moduleSignature).isPresent());
-
-    Assert.assertEquals(1, view_9.getModuleClasses(moduleSignature).size());
-
-    Assert.assertEquals(
-        "de.upb.swt.multirelease.Utility",
-        view_9.getModuleClasses(moduleSignature).stream()
-            .findAny()
-            .get()
-            .getType()
-            .getFullyQualifiedName());
-
-    // for java 9
-    Assert.assertEquals(
-        "/META-INF/versions/9/de/upb/swt/multirelease/Utility.class",
-        view_9.getClass(utilityModule).get().getClassSource().getSourcePath().toString());
-    // different class will be returned if no module is specified
-    Assert.assertEquals(
-        "/de/upb/swt/multirelease/Utility.class",
-        view_9.getClass(utilityNoModule).get().getClassSource().getSourcePath().toString());
-    Assert.assertEquals(
-        "/de/upb/swt/multirelease/Main.class",
-        view_9.getClass(classType2).get().getClassSource().getSourcePath().toString());
-    // assert that method is correctly resolved to base
-    Assert.assertTrue(
-        view_9
-            .getClass(utilityModule)
-            .get()
-            .getMethod(
-                getIdentifierFactory()
-                    .getMethodSubSignature(
-                        "printVersion",
-                        getIdentifierFactory().getType("void"),
-                        Collections.emptyList()))
-            .get()
-            .getBody()
-            .toString()
-            .contains("java 9"));
-
-    // for java 8
-    Assert.assertEquals(
-        "/de/upb/swt/multirelease/Utility.class",
-        view_8.getClass(utilityNoModule).get().getClassSource().getSourcePath().toString());
-    assertFalse(view_8.getClass(utilityModule).isPresent());
-    Assert.assertEquals(
-        "/de/upb/swt/multirelease/Main.class",
-        view_8.getClass(classType2).get().getClassSource().getSourcePath().toString());
-    // assert that method is correctly resolved to base
-    Assert.assertTrue(
-        view_8
-            .getClass(utilityNoModule)
-            .get()
-            .getMethod(
-                getIdentifierFactory()
-                    .getMethodSubSignature(
-                        "printVersion",
-                        getIdentifierFactory().getType("void"),
-                        Collections.emptyList()))
-            .get()
-            .getBody()
-            .toString()
-            .contains("java 8"));
-  }
-
-  @Test
-  public void testApk() {
-    PathBasedAnalysisInputLocation pathBasedNamespace =
-        PathBasedAnalysisInputLocation.create(apk, null);
-    final ClassType mainClass =
-        getIdentifierFactory().getClassType("de.upb.futuresoot.fields.MainActivity");
-    testClassReceival(pathBasedNamespace, Collections.singletonList(mainClass), 1392);
-  }
 
   @Test
   public void testSingleClass() {
